+++ Cargo.toml
@@ -64,7 +64,7 @@ protobuf-codec = [
   "txn_types/protobuf-codec",
   "grpcio-health/protobuf-codec",
   "file_system/protobuf-codec",
-  "req_cpu/protobuf-codec",
+  "resource_metering/protobuf-codec",
   "protobuf/bytes",
 ]
 prost-codec = [
@@ -99,7 +99,7 @@ prost-codec = [
   "txn_types/prost-codec",
   "grpcio-health/prost-codec",
   "file_system/prost-codec",
-  "req_cpu/prost-codec",
+  "resource_metering/prost-codec",
 ]
 testexport = ["raftstore/testexport"]
 test-engines-rocksdb = [
@@ -222,7 +222,7 @@ url = "2"
 uuid = { version = "0.8.1", features = ["serde", "v4"] }
 walkdir = "2"
 yatp = { git = "https://github.com/tikv/yatp.git", branch = "master" }
-req_cpu = { path = "components/req_cpu" }
+resource_metering = { path = "components/resource_metering" }
 
 [dev-dependencies]
 panic_hook = { path = "components/panic_hook" }
@@ -257,6 +257,9 @@ procinfo = { git = "https://github.com/tikv/procinfo-rs", rev = "5125fc1a69496b7
 # After the PR to kvproto is merged, remember to comment this out and change the rev for kvproto dependency.
 # [patch.'https://github.com/pingcap/kvproto']
 # kvproto = {git = "https://github.com/your_github_id/kvproto", branch="your_branch"}
+
+[patch.'https://github.com/pingcap/kvproto']
+kvproto = { git = "https://github.com/zhongzc/kvproto", branch = "pr/773" }
 
 [workspace]
 # See https://github.com/rust-lang/rfcs/blob/master/text/2957-cargo-features2.md
@@ -316,7 +319,7 @@ members = [
   "components/collections",
   "components/coprocessor_plugin_api",
   "components/test_coprocessor_plugin/example_plugin",
-  "components/req_cpu",
+  "components/resource_metering",
 ]
 default-members = ["cmd/tikv-server", "cmd/tikv-ctl"]
