+++ components/resource_metering/src/cpu/collector/mod.rs
@@ -1,11 +1,11 @@
 // Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::RequestCpuRecords;
+use crate::cpu::CpuRecords;
 
 use std::sync::Arc;
 
 pub trait Collector: Send {
-    fn collect(&self, records: Arc<RequestCpuRecords>);
+    fn collect(&self, records: Arc<CpuRecords>);
 }
 
 #[derive(Copy, Clone, Default, Debug, Eq, PartialEq, Hash)]
