+++ components/resource_metering/src/cpu/future_ext.rs
