+++ components/server/Cargo.toml
@@ -113,7 +113,7 @@ toml = "0.5"
 txn_types = { path = "../txn_types", default-features = false }
 vlog = "0.1.4"
 yatp = { git = "https://github.com/tikv/yatp.git", branch = "master" }
-req_cpu = { path = "../req_cpu" }
+resource_metering = { path = "../resource_metering" }
 
 [target.'cfg(unix)'.dependencies]
 signal = "0.6"
